chat_service/requirements_prod.txt
Django==3.2.12
redis==3.5.3
channels==3.0.3
channels-redis==3.2.0
requests==2.28.2
python-decouple==3.4
pymysql==1.0.2
mysql.connector==2.2.9